Move testing utilities into separate folder
