+++ src/application/Layouts/__tests__/FullWidthLayout.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { renderWithApolloClient } from '../../renderWithApolloClient';
+import { renderWithApolloClient } from '../../utilities/testing/renderWithApolloClient';
 import { FullWidthLayout } from '../FullWidthLayout';
 
 describe('<FullWidthLayout />', () => {
