+++ src/application/Queries/__tests__/QueryViewer.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { within } from '@testing-library/react';
 import user from '@testing-library/user-event';
 import stringifyObject from "stringify-object";
-import { renderWithApolloClient } from '../../renderWithApolloClient';
+import { renderWithApolloClient } from '../../utilities/testing/renderWithApolloClient';
 import { QueryViewer } from '../QueryViewer';
 
 describe('<QueryViewer />', () => {
