+++ src/application/utilities/testing/fakeGraphQL.ts
