+++ src/application/utilities/testing/renderWithApolloClient.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { ApolloProvider } from "@apollo/client";
 import { ThemeProvider } from "emotion-theming";
-import { client } from './index';
-import { themes, ColorTheme } from './theme';
+import { client } from '../../index';
+import { themes, ColorTheme } from '../../theme';
 
 export const renderWithApolloClient = (
   ui,
